Moved Command and Task out of cli.go
